+++ src/info/bitcoinsv.js
@@ -4,7 +4,6 @@ import { type EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import type { EngineCurrencyInfo } from '../engine/currencyEngine.js'
 import type { BcoinCurrencyInfo } from '../utils/bcoinExtender/bcoinExtender.js'
-import { imageServerUrl } from './constants.js'
 
 const bcoinInfo: BcoinCurrencyInfo = {
   type: 'bitcoinsv',
@@ -78,13 +77,10 @@ const currencyInfo: EdgeCurrencyInfo = {
   metaTokens: [],
 
   // Explorers:
-  blockExplorer: 'https://blockchair.com/bitcoin-sv/block/%s',
-  addressExplorer: 'https://blockchair.com/bitcoin-sv/address/%s',
-  transactionExplorer: 'https://blockchair.com/bitcoin-sv/transaction/%s',
-
-  // Images:
-  symbolImage: `${imageServerUrl}/bitcoinsv-logo-solo-64.png`,
-  symbolImageDarkMono: `${imageServerUrl}/bitcoinsv-logo-solo-64.png`
+  blockExplorer: 'https://blockchair.com/bitcoin-sv/block/%s?from=edgeapp',
+  addressExplorer: 'https://blockchair.com/bitcoin-sv/address/%s?from=edgeapp',
+  transactionExplorer:
+    'https://blockchair.com/bitcoin-sv/transaction/%s?from=edgeapp'
 }
 
 export const bitcoinsv = { bcoinInfo, engineInfo, currencyInfo }