+++ src/info/constants.js
@@ -1,7 +1,6 @@
 // @flow
 
-export const imageServerUrl = 'https://developer.airbitz.co/content'
-export const InfoServer = 'https://info1.edgesecure.co:8444/v1'
+export const InfoServer = 'https://info1.edge.app/v1'
 
 export const FixCurrencyCode = (currencyCode: string): string => {
   switch (currencyCode) {
@@ -9,6 +8,8 @@ export const FixCurrencyCode = (currencyCode: string): string => {
       return 'BC1'
     case 'DGB':
       return 'DGB1'
+    case 'FIRO':
+      return 'XZC'
     default:
       return currencyCode
   }