+++ src/info/dash.js
@@ -4,7 +4,6 @@ import { type EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import type { EngineCurrencyInfo } from '../engine/currencyEngine.js'
 import type { BcoinCurrencyInfo } from '../utils/bcoinExtender/bcoinExtender.js'
-import { imageServerUrl } from './constants.js'
 
 const bcoinInfo: BcoinCurrencyInfo = {
   type: 'dash',
@@ -71,13 +70,9 @@ const currencyInfo: EdgeCurrencyInfo = {
   metaTokens: [],
 
   // Explorers:
-  blockExplorer: 'https://blockchair.com/dash/block/%s',
-  addressExplorer: 'https://blockchair.com/dash/address/%s',
-  transactionExplorer: 'https://blockchair.com/dash/transaction/%s',
-
-  // Images:
-  symbolImage: `${imageServerUrl}/dash-logo-solo-64.png`,
-  symbolImageDarkMono: `${imageServerUrl}/dash-logo-solo-64.png`
+  blockExplorer: 'https://blockchair.com/dash/block/%s?from=edgeapp',
+  addressExplorer: 'https://blockchair.com/dash/address/%s?from=edgeapp',
+  transactionExplorer: 'https://blockchair.com/dash/transaction/%s?from=edgeapp'
 }
 
 export const dash = { bcoinInfo, engineInfo, currencyInfo }