+++ src/info/dogecoin.js
@@ -4,7 +4,6 @@ import { type EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import type { EngineCurrencyInfo } from '../engine/currencyEngine.js'
 import type { BcoinCurrencyInfo } from '../utils/bcoinExtender/bcoinExtender.js'
-import { imageServerUrl } from './constants.js'
 
 const bcoinInfo: BcoinCurrencyInfo = {
   type: 'dogecoin',
@@ -32,13 +31,14 @@ const engineInfo: EngineCurrencyInfo = {
   feeUpdateInterval: 10000,
   customFeeSettings: ['satPerByte'],
   simpleFeeSettings: {
-    highFee: '1000',
-    lowFee: '100',
-    standardFeeLow: '500',
-    standardFeeHigh: '750',
+    highFee: '526316',
+    lowFee: '526316',
+    standardFeeLow: '526316',
+    standardFeeHigh: '526316',
     standardFeeLowAmount: '2000000000',
     standardFeeHighAmount: '98100000000'
   },
+  minRelay: 100000000,
   timestampFromHeader(header: Buffer): number {
     if (header.length < 80) {
       throw new Error(`Cannot interpret block header ${header.toString('hex')}`)
@@ -67,13 +67,10 @@ const currencyInfo: EdgeCurrencyInfo = {
   metaTokens: [],
 
   // Explorers:
-  blockExplorer: 'https://blockchair.com/dogecoin/block/%s',
-  addressExplorer: 'https://blockchair.com/dogecoin/address/%s',
-  transactionExplorer: 'https://blockchair.com/dogecoin/transaction/%s',
-
-  // Images:
-  symbolImage: `${imageServerUrl}/dogecoin-logo-solo-64.png`,
-  symbolImageDarkMono: `${imageServerUrl}/dogecoin-logo-solo-64.png`
+  blockExplorer: 'https://blockchair.com/dogecoin/block/%s?from=edgeapp',
+  addressExplorer: 'https://blockchair.com/dogecoin/address/%s?from=edgeapp',
+  transactionExplorer:
+    'https://blockchair.com/dogecoin/transaction/%s?from=edgeapp'
 }
 
 export const dogecoin = { bcoinInfo, engineInfo, currencyInfo }