+++ src/info/groestlcoin.js
@@ -7,7 +7,6 @@ import { type EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import type { EngineCurrencyInfo } from '../engine/currencyEngine.js'
 import type { BcoinCurrencyInfo } from '../utils/bcoinExtender/bcoinExtender.js'
-import { imageServerUrl } from './constants.js'
 
 const isBech32 = address => {
   try {
@@ -117,13 +116,11 @@ const currencyInfo: EdgeCurrencyInfo = {
   metaTokens: [],
 
   // Explorers:
-  blockExplorer: 'https://blockchair.com/groestlcoin/block/%s',
-  addressExplorer: 'https://blockchair.com/groestlcoin/address/%s',
-  transactionExplorer: 'https://blockchair.com/groestlcoin/transaction/%s',
-
-  // Images:
-  symbolImage: `${imageServerUrl}/groestlcoin-logo-solo-64.png`,
-  symbolImageDarkMono: `${imageServerUrl}/groestlcoin-logo-solo-64.png`
+  blockExplorer: 'https://blockchair.com/groestlcoin/block/%s?from=edgeapp',
+  addressExplorer:
+    'https://blockchair.com/groestlcoin/address/%s?from=edgeapp?from=edgeapp',
+  transactionExplorer:
+    'https://blockchair.com/groestlcoin/transaction/%s?from=edgeapp'
 }
 
 export const groestlcoin = { bcoinInfo, engineInfo, currencyInfo }