+++ src/info/zcoin.js
@@ -4,7 +4,6 @@ import { type EdgeCurrencyInfo } from 'edge-core-js/types'
 
 import type { EngineCurrencyInfo } from '../engine/currencyEngine.js'
 import type { BcoinCurrencyInfo } from '../utils/bcoinExtender/bcoinExtender.js'
-import { imageServerUrl } from './constants.js'
 
 const bcoinInfo: BcoinCurrencyInfo = {
   type: 'zcoin',
@@ -18,6 +17,7 @@ const bcoinInfo: BcoinCurrencyInfo = {
     xprivkey58: 'xprv',
     coinType: 136
   },
+  uriPrefix: 'firo',
   addressPrefix: {
     pubkeyhash: 0x52,
     scripthash: 0x7
@@ -26,7 +26,7 @@ const bcoinInfo: BcoinCurrencyInfo = {
 
 const engineInfo: EngineCurrencyInfo = {
   network: 'zcoin',
-  currencyCode: 'XZC',
+  currencyCode: 'FIRO',
   gapLimit: 10,
   defaultFee: 1000,
   feeUpdateInterval: 60000,
@@ -49,12 +49,12 @@ const engineInfo: EngineCurrencyInfo = {
 
 const currencyInfo: EdgeCurrencyInfo = {
   // Basic currency information:
-  currencyCode: 'XZC',
-  displayName: 'Zcoin',
+  currencyCode: 'FIRO',
+  displayName: 'Firo',
   pluginId: 'zcoin',
   denominations: [
-    { name: 'XZC', multiplier: '100000000', symbol: 'Ƶ' },
-    { name: 'mXZC', multiplier: '100000', symbol: 'mƵ' }
+    { name: 'FIRO', multiplier: '100000000', symbol: 'ƒ' },
+    { name: 'mFIRO', multiplier: '100000', symbol: 'mƒ' }
   ],
   walletType: 'wallet:zcoin',
 
@@ -76,11 +76,7 @@ const currencyInfo: EdgeCurrencyInfo = {
   // Explorers:
   addressExplorer: 'https://insight.zcoin.io/address/%s',
   blockExplorer: 'https://insight.zcoin.io/block/%s',
-  transactionExplorer: 'https://insight.zcoin.io/tx/%s',
-
-  // Images:
-  symbolImage: `${imageServerUrl}/zcoin-logo-solo-64.png`,
-  symbolImageDarkMono: `${imageServerUrl}/zcoin-logo-solo-64.png`
+  transactionExplorer: 'https://insight.zcoin.io/tx/%s'
 }
 
 export const zcoin = { bcoinInfo, engineInfo, currencyInfo }